backend/requirements.txt
# Laoshi Coach Backend Dependencies

# Web Framework for API
Flask>=3.0.0
Flask-RESTful==0.3.10

# Database
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.46
redis>=7.2.0

# Password hashing and auth
passlib>=1.7.4
Flask-JWT-Extended>=4.7.1

# CORS Support
flask-cors>=4.0.0

# AI layer
openai-agents>=0.9.1
litellm>=1.81.13
mem0ai>=1.0.3
asyncio>=4.0.0

# Development dependencies
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0

# Security & Rate Limiting
cryptography>=42.0.0
Flask-Limiter>=3.5.0

